requirements.txt
flask==3.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
python-dotenv==1.0.0
requests==2.31.0
